requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
langchain-google-community
google-search-results
tavily-python
langchain-huggingface
langgraph